+++ src/main/java/nextstep/app/SecurityConfig.java
@@ -13,9 +13,9 @@ import nextstep.security.authorization.manager.AuthorityAuthorizationManager;
 import nextstep.security.authorization.manager.AuthorizationManager;
 import nextstep.security.authorization.manager.RequestMatcherDelegatingAuthorizationManager;
 import nextstep.security.authorization.manager.SecuredAuthorizationManager;
-import nextstep.security.authorization.matcher.AnyRequestMatcher;
-import nextstep.security.authorization.matcher.MvcRequestMatcher;
-import nextstep.security.authorization.matcher.RequestMatcherEntry;
+import nextstep.security.authorization.access.AnyRequestMatcher;
+import nextstep.security.authorization.access.MvcRequestMatcher;
+import nextstep.security.authorization.access.RequestMatcherEntry;
 import nextstep.security.config.DefaultSecurityFilterChain;
 import nextstep.security.config.DelegatingFilterProxy;
 import nextstep.security.config.FilterChainProxy;
