+++ src/main/java/nextstep/security/authorization/manager/RequestMatcherDelegatingAuthorizationManager.java
@@ -3,7 +3,7 @@ package nextstep.security.authorization.manager;
 import jakarta.servlet.http.HttpServletRequest;
 import nextstep.security.authentication.Authentication;
 import nextstep.security.authorization.AuthorizationDecision;
-import nextstep.security.authorization.matcher.RequestMatcherEntry;
+import nextstep.security.authorization.access.RequestMatcherEntry;
 
 import java.util.List;
